Classifier: Refactor the beforeunload warning Move the `onbeforeunload` warning from `RootStore` down to `StepHistory`, where it observes `checkForProgress` for changes. Add a reset on `stepHistory.finish()`, so that the warning is removed when we press Done or Done & Talk.
