+++ packages/lib-classifier/src/store/RootStore.js
@@ -1,7 +1,5 @@
-import counterpart from 'counterpart'
-import { autorun, configure } from 'mobx'
+import { configure } from 'mobx'
 import {
-  addDisposer,
   addMiddleware,
   getEnv,
   onAction,
@@ -23,18 +21,9 @@ import WorkflowStore from './WorkflowStore'
 import WorkflowStepStore from './WorkflowStepStore'
 import UserProjectPreferencesStore from './UserProjectPreferencesStore'
 
-import en from './locales/en'
-
 // Isolate mobx globals. 
 // See: https://github.com/mobxjs/mobx/blob/72d06f8cd2519ce4dbfb807bc13556ca35866690/docs/configuration.md#isolateglobalstate-boolean
 configure({ isolateGlobalState: true })
-
-counterpart.registerTranslations('en', en)
-
-function beforeUnloadListener(event) {
-  event.preventDefault()
-  return event.returnValue = counterpart("RootStore.unloadWarning")
-}
 
 const RootStore = types
   .model('RootStore', {
@@ -60,19 +49,6 @@ const RootStore = types
 
   .actions(self => {
     // Private methods
-
-    /**
-      Add or remove a beforeunload listener whenever self.subjects.active?.stepHistory.checkForProgress changes.
-    */
-    function _observeWorkInProgress() {
-      const subject = tryReference(() => self.subjects.active)
-      const { addEventListener, removeEventListener } = window
-      if (subject?.stepHistory.checkForProgress) {
-        addEventListener && addEventListener("beforeunload", beforeUnloadListener, {capture: true});
-      } else {
-        removeEventListener && removeEventListener("beforeunload", beforeUnloadListener, {capture: true});
-      }
-    }
 
     function _addMiddleware(call, next, abort) {
       if (call.name === 'setActiveSubject') {
@@ -113,8 +89,6 @@ const RootStore = types
 
     // Public actions
     function afterCreate () {
-      const subjectAnnotationsDisposer = autorun(_observeWorkInProgress)
-      addDisposer(self, subjectAnnotationsDisposer)
       addMiddleware(self, _addMiddleware)
       onAction(self, _onAction)
       onPatch(self, _onPatch)
