+++ packages/lib-classifier/src/store/SubjectStore/Subject/StepHistory/locales/en.json
@@ -1,5 +1,5 @@
 {
-  "RootStore": {
+  "StepHistory": {
     "unloadWarning": "You have unsaved work. Are you sure you want to leave?"
   }
 }
